requirements.txt
# Manipulação de Dados e Persistência
pandas>=2.0.0
openpyxl>=3.1.0  # Para ler/gravar Excel
SQLAlchemy>=2.0.0

# Geração de Relatórios
fpdf2>=2.7.0

# Tipagem Estrita e Variáveis de Ambiente
python-dotenv>=1.0.0
streamlit>=1.30.0
yfinance>=0.2.0
plotly>=6.6.0
gspread>=6.2.1